get branch to build
SVN revision 5111

+++ src/io/trove/pvfs2-storage.h
@@ -36,7 +36,7 @@ typedef struct PVFS_ds_keyval_s PVFS_ds_keyval;
 
 struct PVFS_vtag_s
 {
-    unsigned long version;
+    uint32_t version;
 };
 typedef struct PVFS_vtag_s PVFS_vtag;
